Add the environment as INIManager variable
In order to be able to access to the environment from every module, a copy of the env. is staged in the INIManager module. We need to access to the environment to know the position of each object and so to be able to target the object to gaze. Before this modification, the BehaviorRelizer had to be connect to the Environment module to read the positions. Know this connection it is not necessary.