Gaze model adjustment
The position of the head agents is not take from the CharacterManager hashmap but directly from the MPEG4 of that character that can be found as children of the environment treenode. 
In this way to gaze an agent, the attribut target in the fml or bml should be contain just the name of the agent, all uppercase (i.e. target="CAMILLE").